Creating CopyTextField component and adding it to showCompletedUpload and Shaere Modal
